cloud-functions/meeting_minutes/requirements.txt
flask
google-cloud-storage
firebase-admin
PyMySQL
sqlalchemy
requests